sensor 5, keypad
Added 5th sensor. Also did a quick test of the numeric keypad and got
it working, though if the sensors are running it tends to eff with the
button detection - too many interrupts! Moved main loop to the end of
the program for readability. This is not a refined commit - just
tossing it up as is in case of fire.