Keypad, sorting out LEDs
In this commit, I add code that enables someone to lock the room via a
numeric keypad (need to add some error checking still to prevent buffer
overflow).

I also changed the LED to an analog write to limit the amount of
current drawn.

The button has been changed so that it unlocks the room instead of
locking it.

Lots of this code is slapped together, so it needs to be prettied up.